Use pico runtime supplied with gnat_arm_elf crate
Now that the pico runtimes are available in the alire supplied
gnat_arm_elf crate, we pull the ravenscar capability from there. We are
still using a fork of the pico bsp that does not conflict with
capabilities supplied in the pico runtimes in gnat_arm_elf. There is no
version of the pico bsp that doesn't supply conflicting items like
interrupts besides the fork specified here.