Add Analyze and Safe build targets for Pico
The Analyze target is meant to be run with GNAT SAS on a CI server,
since it forces deep mode analysis. The Safe target is meant to provide
a "safe image" compile that has runtime checking and optimization turned
off. This can work around compile bugs or runtime check bugs in a last
resort image.